Add `TypeCodes::Map` and `Values::Map` (#21)
This PR builds on #18 to introduce a `Map` to `TypeCodes` and `Value`. The key types and value types for a map are both homogeneous, and keys can only be `MapKey` types (i32, i64, bytes, string). Values can be any `Value` type.

Note that, because we didn't use the primitive/complex type hierarchy in #17, I ended up having to duplicate some logic for `MapKey` that already exists in Value. Notably, the read/write stuff.