fix: ensure that payload has a determinitic order
fixes https://github.com/imprint-serde/imprint/issues/26

Ensures that the payload has a deterministic order by sorting the
payload according to the entries in the directory. The assumption
is that the payloads or both records are already sorted so that
the result can be merge-sorted.

In general r1.merge(r2) is not equal to r2.merge(r1), since if
the records contain same fields the payload of the first record
is written first to the result. The payload of the second record
might or might not be written to the result depending on the
merge option filter_duplicate_payloads.